clean up helper code, add withKeyItemId matcher for KeyListAdapter
